feat: implement file preview functionality with image support and size limit
