+++ apps/meteor/client/lib/constants.ts
@@ -2,3 +2,4 @@ export const USER_STATUS_TEXT_MAX_LENGTH = 120;
 export const BIO_TEXT_MAX_LENGTH = 260;
 export const VIDEOCONF_STACK_MAX_USERS = 6;
 export const NAVIGATION_REGION_ID = 'navigation-region';
+export const MAX_FILE_SIZE_PREVIEW = 10485760; // 10MB
